Security fix: restrict public key verification messages to 31 bytes
Previously, blindly signing up to 32 bytes of arbitrary data was allowed, which
could be abused for signing messages following the Nimiq standard for signed
messages, which are 32 byte Sha256 hashes. This commit fixes that.
In the future, we'll further consider disallowing signing of arbitrary messages
altogether even below the 31 byte threshold, by prefixing and hashing the passed
message or removing the feature.