Security hardening: Strict data length validations
- Check that apdu length matches the advertised length. This was not checked for
  previously.
- Consequently check that any data read from passed data does not exceed input
  data length. Previously such checks were missing in a few places like reading
  of bip32 paths or transaction parsing, apart from transaction extra data.
- Check that after parsing input data there's no leftover data. This was not
  checked for previously in some places like public key requests and transaction
  parsing, apart from transaction extra data.

The check of exact data lengths is especially also in preparation of adding a
transaction version byte in the next commit for distinguishing legacy trans-
actions from Albatross transactions. By checking for exact data lengths, we can
detect if the version byte is missing, and this way reject supporting old ver-
sions of the client api. Instead, devs using the client api, should update the
api accordingly. On the other hand, new versions of the api can detect which
version of the Ledger app they're communicating to, and accordingly send the
transaction version byte or not, to keep supporting old versions of the Ledger
app not updated by the user yet.