Introduce a transaction version byte
Add a transaction version byte to transaction signing requests to identify
legacy or Albatross transactions, or future transactions formats. This commit
only adds handling of this version byte, no support for Albatross yet.

Notably, by checking for exact data lengths as of the previous commit, we can
detect if the version byte is missing, and this way reject supporting old ver-
sions of the client api. Instead, devs using the client api, should update the
api accordingly. On the other hand, new versions of the api can detect which
version of the Ledger app they're communicating to, and accordingly send the
transaction version byte or not, to keep supporting old versions of the Ledger
app not updated by the user yet.