Automatically create staker signature proof
For incoming staking transactions which are meant to include a staker signature
proof in their recipient data but only include the empty default proof, we
replace that empty signature proof with an actually signed staker proof. For
this, the same ledger account is used as staker and transaction sender, which is
the most common case for regular users. This way, no separate signature request
needs to be sent to the ledger to first create the staker signature proof, but
both signatures are created in a single request for better UX. For advanced
users, usage of a different staker than the transaction sender is still
supported by providing a ready pre-signed signature proof in the request instead
of an empty signature proof.